Add solution for desired outcome
My first solution is making the card class into flex container with wrap and the margin-right of header to auto so the cancle button goes to the right most place. Although it looks very similar, the header and cancle button did not align vertically so I made adjustments according to the solution. My initial solution do not need margin bottom cos it used with gap